requirements.txt
aistudio-sdk==0.3.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyconfig==0.14.0
anyio==4.12.1
astor==0.8.1
attrs==25.4.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
easyocr==1.7.2
editdistance==0.8.1
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
munch==4.0.0
narwhals==2.16.0
natsort==8.4.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.92
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.8.1
paddlepaddle-gpu==2.6.1
paddlex==3.4.2
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
prettytable==3.17.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.0
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.4.0
pyperclip==1.11.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
scikit-image==0.26.0
scipy==1.17.0
sentry-sdk==2.52.0
setuptools==70.2.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8.3
streamlit==1.54.0
sympy==1.13.1
tenacity==9.1.4
tensorboardX==2.6.4
termcolor==3.3.0
tifffile==2026.1.28
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.3
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
wandb==0.24.2
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.5
zope.event==6.1
zope.interface==8.2
